% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nasalPkgFunctions.R
\name{parse_taxonomy_strain}
\alias{parse_taxonomy_strain}
\title{Parse Greengenes-style taxonomy strings including a Strain rank}
\usage{
parse_taxonomy_strain(char.vec)
}
\arguments{
\item{char.vec}{Character vector of taxonomy rank strings (already split by rank),
typically provided by \code{phyloseq::import_biom()}.}
}
\value{
A named character vector with names:
\code{Kingdom, Phylum, Class, Order, Family, Genus, Species, Strain}.
}
\description{
Helper parse function for \code{phyloseq::import_biom()} that converts a split
Greengenes taxonomy character vector into a named vector of taxonomic ranks.
In addition to standard Greengenes ranks, this parser includes an extra
\code{"Strain"} rank.
}
\details{
The input \code{char.vec} is assumed to already be split into rank entries
(as \code{import_biom()} does for Greengenes-style taxonomy). Each entry is
expected to have a 3-character prefix such as \code{"k__"}, \code{"p__"}, etc.,
which is removed.
}
\examples{
\dontrun{
x <- c("k__Bacteria","p__Firmicutes","c__Bacilli","o__Lactobacillales",
       "f__Streptococcaceae","g__Streptococcus","s__aureus","t__USA300")
parse_taxonomy_strain(x)
}

}
\keyword{internal}
